:bug: Remove cluster label from existing lb on delete
When a cluster is deleted, the load balancer is usually deleted.
However, if the user specified the name of an existing load balancer in
the spec, the expectated behavior would be that the load balancer of the
user is not deleted when a cluster is deleted.

This commit implements that behavior and also removes the cluster label
from the load balancer.

It also improves the conditions and events for load balancer deletion.

A problem with pointers is fixed which caused an issue in the fake
hcloud client, where we implicitly updated the load balancers with the
pointer, instead of using the actual API call.

In order to fix this, the API fake behavior is fixed and labels are also
updated, according to the behavior of the actual hcloud-go client